refactor(contracts/lib): Update file structure and names
