--- libs/ts/contracts/src/clients/index.ts
@@ -1,3 +1,0 @@
-export * from './aggregated-data-feed-store';
-export * from './cl-aggregator-adapter';
-export * from './contract-consumer';
